1. Adding an ability to create a bundle of policy flows. 2. Adding an ability to show/edit/delete one policy flow.

+++ app/Model/Request.php
@@ -97,7 +97,9 @@ class Request extends Model
             $request_overview->delete();
         }
 
-        $this->file->delete();
+        if ($this->file) {
+            $this->file->delete();
+        }
 
         return parent::delete();
     }
